pyproject.toml
[project]
name = "delta-scoreboard"
version = "0.1.0"
description = "A Dash app for workshop leaderboard hosted on Databricks"
authors = [
    {name = "Workshop Team", email = "workshop@example.com"},
]
dependencies = [
    "dash>=2.17.0",
    "dash-bootstrap-components>=1.5.0",
    "pandas>=2.0.0",
    "databricks-sdk>=0.20.0",
    "databricks-sql-connector>=3.0.0",
    "plotly>=5.17.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "gunicorn>=21.0.0",
    "requests>=2.31.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = "MIT"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
py-modules = ["main", "main_demo", "config", "databricks_utils", "demo_mode"]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 